feat: add types file and nested props
