+++ src/custom-jsx-library/jsx-runtime.ts
@@ -1,19 +1,5 @@
-import { createElement, Key } from "./core";
-
-export interface JSXProps {
-  children?: any;
-  [key: string]: any;
-}
-
-export interface JSXNode {
-  type: ElementType;
-  config: JSXProps;
-  key: Key;
-}
-export type ElementType = string | Function;
-export type JSXElement = JSXNode | string | number | boolean | null | undefined;
-
-export type JSX = (type: ElementType, config: JSXProps, key?: Key) => any;
+import { createElement } from "./core";
+import { JSX } from "./types";
 
 export const jsx: JSX = (type, config, _key) => {
   const props = { ...config };
@@ -23,9 +9,6 @@ export const jsx: JSX = (type, config, _key) => {
 };
 
 export const jsxs: JSX = (type, config, key) => {
-  return jsx(type, config, key);
-};
-export const jsxDEV: JSX = (type, config, key) => {
   return jsx(type, config, key);
 };
